fix: update docker-compose to use image built from Dockerfile
The image pulled and used in devstack by default continues to be the container built from our legacy ansible configuration. This means that developers aren't testing and developing against the same image deployed in Stage and Production.

Additionally, the legacy image is running older versions of Node and npm (Node 16 and npm 8.x, instead of Node 18 and npm 9.9.x).

This PR updates devstack to use the image built from the Credentials dockerfile, which is built after every merge to the `master` branch.