fix: Fix MFE repo URLs; better error message on bad format; tighter regex (#44)
Two frontend repos were missing the `.git` suffix that is required for
the regex that maps repo URLs to directory names. That's fixed now.

But also, `make dev.reset-repos` would print two lines of `The [] repo is
not cloned. Skipping.` -- hard to understand that that meant the regex
failed to match. Now the match failure is explicitly handled for checkout,
clone, reset, and status and a useful message is printed.

Also, tightens up regex to anchor the end.