feat: repo.sh - shallow cloning still tracks remote branches.
When setting SHALLOW_CLONE=1 setting for the `make dev.clone` target,
developers are left with clones which cannot checkout remote branches
for development. They are forced to search stackoverflow and ask AI bots
how to fix this problem and discover an obscure command to fix their
local clone.

This PR just captures that work so others don't suffer.