fix: Many fixes for enterprise IDAs
It seems like the various enterprise-specific devstack settings are a
mix of legacy stuff combined with failed attempts to port repo-specific
devstack configs into this central devstack repo. As a result, the
various hostnames and paths are completely broken. This hasn't impacted
us because we normally don't use enterprise inside of devstack, but this
is will hopefully be our future.